+++ android/app/src/test/java/app/chompass/MainActivityDebugExtrasGateTest.kt
@@ -45,6 +45,7 @@ class MainActivityDebugExtrasGateTest {
             putExtra("run_entry_benchmark", true)
             putExtra("benchmark_count", 9)
             putExtra("run_relog_benchmark", true)
+            putExtra("run_water_sip_benchmark", true)
             putExtra("run_ondevice_llm_test", true)
             putExtra("diagnose_health_connect", true)
         }